Enable Vulkan validation layers on Linux
When vkDestroyDebugReportCallbackEXT() is called in Backend::~Backend()
and Vulkan validation layers are in use, the variable unique_id_mapping
is referenced internally[1]. However, the variable is global[2] and it's
destructed before Backend::~Backend(), which causes a use-after-free
issue. The issue was unnoticed on Windows, but we observed the crash at
exit on Linux.

[1] https://chromium.googlesource.com/external/github.com/KhronosGroup/Vulkan-ValidationLayers/+/9fba37afae13a11bd49ae942bf82e5bf1098e381/layers/generated/layer_chassis_dispatch.cpp#4961
[2] https://chromium.googlesource.com/external/github.com/KhronosGroup/Vulkan-ValidationLayers/+/9fba37afae13a11bd49ae942bf82e5bf1098e381/layers/generated/chassis.cpp#40

Bug: dawn:150
Change-Id: I505373a88ef9795243dd18da9785fb49d253e498
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/13787
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>