docs: Add note about dawn_perf_tests without a Chromium checkout
Building from a Chromium checkout with all the optimization flags
is most useful when LTO makes a signficant performance difference.

This patch updates the docs to be clear that testing performance can
still be done without a Chromium checkout. This is the easiest path
for external partners.

Bug: dawn:23
Change-Id: I6ee0f0a94f14f12973d0ef4c62298278f11fb7a0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/14503
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>