Adding PRESUBMIT.py with a bunch of checks turned on
This CL does not retroactively fix violations of the checks, so as new
code is added unrelated failures may appear from older code in the same
file.

Removed the headers= line from CPPLINT.cfg, since it was adding an
extension,.hpp, that we don't use for headers and it is not supported on
all platforms.

BUG=tint:1

Change-Id: I9b2d096cf6c69d1456c857a2fdcd7e7da3fd747b
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/22544
Reviewed-by: David Neto <dneto@google.com>
Reviewed-by: dan sinclair <dsinclair@google.com>